Added `AUDIOFOCUS_NONE` (int 0) as an option for `AndroidAudioFocusGainType`
I'm trying to create a feature in which my app can be played with other apps. When that feature is enabled, I don't want to make other apps paused when starting the audio session (calling `AudioSession.instance.setActive(true)`. This wasn't possible in the original package so I forked it and now added it as an option according to android docs.

https://developer.android.com/reference/android/media/AudioManager#AUDIOFOCUS_NONE